Fixing bug in recent projects open
If one recent project failed to get metadata, they would all fail

The root cause of the bug was a failure to `await` a Promise
inside a `try/catch` block in an async function.

By just returning the Promise instead of `await`-ing it first, the
throw would happen in a different context, and that would
change the semantics of the program.