[npm] Use the exp-latest branch of react-native
This matches the version of react-native that we use in the development version of Exponent. For now, since the packager ends up bundling the files in its own copy of react-native (instead of the copy in the project's own `node_modules`), we need to make sure that XDE's react-native matches the Exponent client's.