Download simulator apps on demand instead of including them in the app
We download different simulator apps based on the version of react-native
that is included in this version of xde. We base that on two things --
    1. The version of react-native described in xde's package.json
    2. The self-reported version of react-native in react-native's package.json

If that isn't already there, we can send an API request to our servers
and then get back a URL to download. This version of xde will take care of that.

Right now, we always return a URL to the one version of the simulator apps
that we have (just the app store version), but in the future this can be
much more flexible, esp. as the packager gets split out, and one version
of xde can work with multiple versions of React Native.