Unbreaking installation of simulator app
@nikki figured this out.

Basically, we had a line looking for the react-native version in the
main xde package.json. Over time, the way we got the path for the simulator
version of the app within the XDE bundle changed, and we didn't use that
data anymore, but the line requesting it was still in the code.

Eventually, the react-native dependency was removed from xde. This caused
the install function to blow up when the version of react-native was requested.

The fix is super simple -- just removing the (now failing) request for the
version data we don't even use.