Fix popover arrow styling.
Summary:
So sorry for flip-flopping on this @_@

Moved the gray border/box shadows out of `Menu` (now deleted) and back into
`Popover` so I could control the z-index of the layers.

The stacking looks like:
1) arrow border (gray)
2) content border (gray border and box shadows)
3) arrow fill (white)

This prevents the content border from cutting a line under the arrow (which is
actually a diagonal box).

Test Plan: Manual testing of XDE.

Reviewers: jesse

Reviewed By: jesse

Subscribers: ccheever

Differential Revision: D2971

fbshipit-source-id: 7df0054