Use electron-rebuild to build node_modules
Summary:
electron-rebuild is a helper for compiling native node_modules so that they match the ABI version of Node & V8 that is packaged with Electron. This is needed because Electron uses ABI 49 while Node is still on ABI 48.

To avoid rebuilding node_modules each time, electron-rebuild internally has some heuristics to compare Electron's ABI version with that of Node running electron-rebuild. But this will never match up since Electron's on 49 and Node 6 is on 48. So instead I made it check if the C++ headers for the current version of Electron have been downloaded, and if so, don't rebuild.

This can cause issues if we install new node_modules with native dependencies under `xde/app`. The fix is to delete `xde/node_modules/electron-rebuild/lib/headers` and run `gulp build` again. Maybe `gulp clean` should take care of this too? Fortunately native modules aren't something we install too often...

Test Plan: Ran `gulp build` which recompiled node_modules the first time but not the second.

Reviewers: jesse

Reviewed By: jesse

Differential Revision: D3008

fbshipit-source-id: d4bb456