Fix passing down IconButton styles so it works with Aphrodite
Added a `styles` prop as suggested by the Aphrodite maintainers. This prop is an array of Aphrodite styles that then get spread and passed into `css()` at the end.

This fixes the styles that were intended to be passed into IconButton's button but weren't getting applied.

Tested by opening XDE and seeing the buttons nicely spread out again.

fbshipit-source-id: 13bae7b